modify: 修改分页模型的数据类型

+++ minion-core/minion-core-mybatis/src/main/java/cn/dhbin/minion/core/mybatis/model/IPageModel.java
@@ -13,14 +13,14 @@ public interface IPageModel<T> {
      *
      * @return 页大小
      */
-    long getSize();
+    Integer getSize();
 
     /**
      * 获取当前页
      *
      * @return 当前页
      */
-    long getCurrent();
+    Integer getCurrent();
 
     /**
      * model -> T
